Initial implementation of accessing external probes
These changes are meant to add other repositories and their
probes to the probe-scraper output.

The users input their repo information in repositories.yaml.
The program is meant to run daily, and each day the scraped files
are added to the previously scraped ones. This means that the
previous day's scrapes need to be available, which can be done by
having a dedicated machine or by using external storage and syncing.

If there is a problem downloading the file or processing it,
an email is sent with the issue.

The probes are outputted in similar format, except that they
do not have _version_ or _node_. Instead they have _date_,
which contains the first and last date we saw that definition.

The files are outputted to
`output_dir/<repo_name>/mobile-metrics/all-probes`.

Eventually we can include whether it should be `main` ping probes
(which indicates addon scalars and histograms) or mobile-metrics,
so this is extensible to future data.

*Summary of Changes*
- Move scrapers to `/scrapers/`, rename existing scraper to
  `moz_central_scraper`
- Add `repositories.yaml`, which indicates external repos to
  scrape
- Add `http_scraper`, which scrapes all files in `repositories.yaml`
- Add separate runner for scraping and storing http-accessed repos
- Add `emailer.py`, which is a simple email API
- Add capability to `transform.py` to merge definitions by date
- Add tests to `transform.py` for this capability