1568485: Add engine-gecko dependencies to repositories.yaml
Even though there is a scraper for mozilla-central, the repositories.yaml
feature for Glean is currently limited to git repositories.  Therefore, this
uses the read-only mirror of the gecko-dev part of mozilla-central on Github.

There was a small code change required to support only looking on a specific
branch of the repo.