Add descriptions, channels to repositories
We currently have no easy way of understanding what application
an individual repository corresponds to. For the purposes of the
Glean dictionary and other tools, we really should. Even
internally inside the probe scraper, having this metadata makes
it easier to understand what data we are collecting for auditing
purposes.

We also add an (optional) channel property, currently limited to
the options described in this document:

https://docs.telemetry.mozilla.org/concepts/channels/channel_normalization.html

Finally, add a deprecated boolean, for repositories which are
deprecated and should probably not be shown by default in most
views.

This PR:

* Adds description (mandatory), channel (optional), deprecated
  (optional) fields to the repository schema
* Adds description to all repositories, in line with the above

Updates

Channel, deprecated fields

Expand repository parser test case