Move Glean to glean-deprecated; add library name
Previously, we were setting this value in MSG as the default
dependency.

In an effort to remove glean (in favor of glean-android and
glean-core), we will move these apps off of the default
dependency.

Once this is done, we can move the default dependency in MSG to
glean-core (alternatively, we can remove the default dependency).