Fix marking whether metrics are in source code or not (#329)
In my previous approach, I was using a commit index (not a timestamp, as
I thought I was),  which is actually *lower* for the latest value. This
would work only for metrics with one history entry: other metrics would
always be marked as not "in source". The right solution is actually
really simple: just see if the metric is present in the latest commit
index ("0") -- if it is, it's in source. If it isn't, it's not. No need
to bother with timestamps at all.

Hopefully fixes #320 for good this time.