bug 1724303 - Reorganize Gecko metrics into its own library.
Metrics and Pings collected in the reusable parts of libxul will now (soon?) be
reported from all applications using the reusable parts of libxul.
(See bug 1670261) Specifically Fenix and Firefox Desktop, though I wouldn't be
surprised to see Focus and Klar (Android) included in the near future.

Though we could split this by channel, it'll trip the probe-scraper up as new
metrics files are added to gecko first in nightly and then ride the trains.