Fix buildhub fetcher to account for v100 and upwards
The `target.version` in buildhub is a string field like `99.0a1`.
Now `"100" < "99"` in string comparison, which is what's used for the
elasticsearch-powered search API.
We therefore failed to have those versions included when asking buildhub
for a list of releases.

Fixing that is a bit more involved.
If we ask for versions from something below 100 and upwards we now
explicitly include everything between 100 and 200.
Yes, that will break on us with v200. At current release rate and if my
math is correct that's gonna be in 2029. I'm ok with not caring about
that now.

Fixes #411