Bug 1764332 - Update bergamot to point to the legacy branch
The probe-scraper requires files to be always available and it will
break if products are removed or not scrape-able. To unblock the current
runs, we need to point the offending product to the legacy branch.
Moreover, since removing a product would cause problems downstream
(schema generator), we remove the expectation for the missing Bergamot
files.