Bug 1769579 - Use the Rally monorepo as a reference
This changes the Rally Core Add-on entry to scrape data from
the Rally monorepo instead of the legacy core-addon. The YAML
files are shared across the two projects. The team intends to
use the same secure environment to process data coming from
both the add-on and the web-platform (they will use the
channel and the version to distinguish them).