Fix how git scraper handles a single commit (#474)
`git log --max-count=1 <ref> -- path/to/file.yaml` returns the most recent commit in the history of `<ref>` to modify the file, but we actually just want to return `<ref>` when `<ref>` is a commit. `--max-count` only produced the correct result when the repo was shallow, and the repo may be unshallowed by checking that a commit is in a given branch when the commit is not the current head of that branch.

In order to achieve the desired output, this PR removes the ineffective `max_count` argument in favor of a new `only_ref` argument.