Bug 1797446 - Don't cc folks as it might break
Bugzilla doesn't allow cc'ing non bugzilla accounts. cc'ing
non bugzilla accounts will make the REST API call fail. I think
we can safely skip cc'ing people as cc will be done automatically
when adding a needinfo and the expectations is that bugs will be
handled with triage once filed, so no cc is needed if we are not
able to needinfo.