Bug 1940030 - Ignore broken_site_report.* reports for dupe-detection
These are initially defined in gecko (though this will change),
but the existing code was only ever sending them from Firefox Desktop.
With recent changes they are redefined in Fenix and submitted from there too.

Because the tables were already generated based on the old schema we can
receive them just fine.
We can thus skip them in the dupe-detection check.

This hack (along with the gecko.* metrics also in that list) can
hopefully go away once m-c changed to define the metrics for
firefox-desktop instead.